Moved find_queue_families back to physical_device
also completed the find_queue_families function
